Move delegation test to its own file
